Collect valid function declarations to do transformations.
Here the validity of a function depends on the types of its parameters. 
Currently, functions which have at least one integeral or enumeration parameter
are legitimate. 

Also record the number of valid transformation instances. 
This number actually depends on the occurences of integeral/enumeration
parameters. For example, we have two instances for the code below:

foo(int p1, char p2)