Fixed some problems about named and un-named parameters.
Clang gives different RangeSize values on named and un-named parameters.
For example, for the first parameter,

  foo(int, int);  -- RangeSize is 4, i.e., "," is counted
  foo(int x, int);  -- RangeSize is 5, i.e., ","is not included

Need to separate the routine depending on the value returned
by ParmVarDecl::getDeclName()

Also, based on the comments in clang/lib/Rewriter/Rewriter.cpp,
we'd better not to rely on the return value of SourceRange.getEnd()
when forwarding search ',' symbol. 

Insteadly, use getStart().