Added another simple transformation --return-void.
This transformation changes the return type of a 
non-void-return-function to void, and replaces all 
of the return statements in the function with "return". 

Note that the transformation is not sound. If a return
value of the transformed function is used in some
expressions, the generated code will be complained by 
testing compilers.