Included local-to-global transformation
This one lifts local decls to global scope and
also renames the lifted variables. 

Note that the transformation is unsound. For example,
it will transform the code below to a bad form:

foo() {
  int x = 1;
  int y = x;
}

Assume "y" is the lifted variable, the code will
be transformed to:

int foo_y = x;

foo() {
  int x = 1;
}