Added another transformation remove-nested-function
This transformation removes a nested function invocation
from its callsite. The transformation creates a temporary
variable with the correct type, assgins the return value
of the selected nested function to the newly created variable,
and then replace the function invocation with this temporary
variable.

Note that this is another unsound transoformation because
it changes the execution semantics of certain foo-loops.
For example:

For example,

  int foo(void)
  {
    int i;
    int x = 1;
    for(i = 0; i < bar(bar(x)); i++)
      ...
  }

will be transformed to:

  int foo(void)
  {
    int i;
    int x = 1;
    int tmp_var = bar(x);
    for(i = 0; i < bar(tmp_var); i++)
      ...
  }

This is wrong because the loop body might 
changed the value of variable x.