Added one more pass replace-callexpr
It replace a CallExpr with a return expression from 
this callee's body. The transformation fires only
if the components of the return expression are one
of the following:
  * global variables
  * parameters
  * constants
If a callee has several return statements, all of them 
will be exercised separately, i.e., the transformation
will choose one for each iteration.

This pass is unsound because the side-effects on globals
and parameters caused by callees are not captured.

Despite the unsoundness, I think this transformation is still 
useful because it could remove quite a few trivial invocations, 
e.g.,

int foo(int p) { return p; }
int bar(void) {
  int x;
  x = foo(...);
}

where foo(...) will be removed.

Anyway, a better inlining pass and copy-propogation pass are 
still needed.