Seems reduce-pointer-level works now.
Basically this pass reduces the indirect level of 
a pointer each time in a quite conservative way. 
It only attempts to handle global/local pointers.

All valid variables are sorted by their indirect levels.
The pass will ensure to first choose a valid variable
with the largest indirect level. This approach could
reduce the complexity of our implementation, because
we don't have to consider the case where the chosen variable
with the largest indirect level would be address-taken.
Variables at non-largest-indirect-level are ineligible
if they: 
  * being address-taken
  * OR being used as LHS in any pointer form, e.g.,
    p, *p(assume *p is of pointer type),
    while the RHS is NOT a UnaryOperator.

If a reduced pointer has an initializer,
the transformation will create a newly valid initializer 
for it. 

If this pointer has global scope, it will be
given the correct initialization value from its pointed-to
object, e.g.,
  int g1 = 1;
  int *p1 = &g1;
  int **p2 = &p1;
Assume p2 is the pointer under reduction, we will have
the following transformed code:
  int *p2 = &g1;

If this pointer has local scope, I just simply remove the
address-taken symbol on the RHS, e.g.,
  int *p2 = p1;

I skipped field pointers, which are not supported by Csmith.
I marked it as future work.