Enabled union-to-struct pass
Change a union declaration to a struct declaration.
For a union declaration which has only scalar fields, i.e., fields
with integer or pointer types, the transformation will set up an
appropriate initializer for a variable which is declared with a union
type under transformed:
  * if originally a pointer field gets initialization value, then
other pointer fields with the same type will get the same initialization
value, and all other fields will be initialized to 0;
  * if originally a integer field gets initialization value, then all
other integer fields will be initialized to the same value, while all
pointer fields will have 0.

For other cases, all initializers will be dropped. (Note that this could
bring in uninitialized local variables.)