Handle a special case where we have empty InitList
For example, we could have code like this:
  union U0 a[][2] = {};

In this case, it's safe to keep the empty initializer.
Sometimes we want to keep this empty initializer because if
"a" is a local var, remove the empty initializer will trigger
compiler errors and hence we can't replace union with struct.
Note that if "a" is defined in a dead block, then we don't
have an UB here. I've seen a case where this dead definition
caused a wrong code bug.