Use a more specific `clang_delta' binary.
If the `creduce' script seems to be installed, use the `clang_delta'
binary that is also installed.  Otherwise, assume that the `creduce'
script is in a build tree and look for the `clang_delta' binary in the
build tree.

A future enhancement would allow one to specify the `clang_delta'
binary to be used, but I don't know if that would be very useful in
practice. In any case, implementing such a thing would require some
sort of parameter-passing to `check_prereqs'.