Put working files `$cfile.bak' and `$cfile.orig' in the cwd.
Previously, the ".bak" and ".orig" versions of the input file ($cfile)
were written to the directory that contains the input file.  Now the
`creduce script is careful to put these working files in the current
working directory.

Note that the filename-parsing doesn't work right if a user insanely
does this:

     creduce test.sh foobar.c/

where `foobar.c' is a symlink to an existing file.  Go figure.  Maybe
in a future life I will fix this.