Tweak the shebang line so `perl -- creduce.in' can work.
This is an experiment toward developer convenience.

When hacking `creduce.in', a developer might not want to re-run Make,
but instead run the modified `creduce.in' script directly.  One can do
that as follows (perhaps wrapped up in a shell alias):

     perl -- creduce.in <arg> <arg>

But to make this work, the shebang (#!) line must contain "perl".
See the `perlrun' man page for the ugly details.

Perl is weird.