Implement more portable shebang magic for `creduce.in'.
Parsing shebang lines is incredibly complicated, and the shebang line
in my previous commit is potentially nonportable because the "#" may
not be treated as a comment:

	#!@PERL@ # -*- perl -*-

For some discussion of this, see:
http://www.in-ulm.de/~mascheck/various/shebang/#comments
http://lists.freebsd.org/pipermail/freebsd-arch/2005-February/003525.html

My new solution is to be simpler.  Use the configure-time substitution
`@perl@' to make Perl happy about running `creduce.in', if that's what
you really want to do:

	perl -- creduce.in

But continue to support `@PERL@' at configure-time, too.