added another pass --replace-array-index-var
if we have ArraySubscriptExprs inside a for loop and these ArraySubscriptExprs
are indexed by a loop inducing var, then replace the indices with constants
from 0 to (sz - 1), where sz is the size of the array. Also, to decrease
the change of generating too much transformations, this pass only applies
to one-dimensional arrays with size less than 15.