Refresh the Autoconf/Automake files.
The main purpose of this is to include recent Perl module checks
in `configure'.

The rest is just "keeping up to date," which is somewhat brought on
by the fact that my primary work computer changed since the last time
I installed/built this stuff.  So I have different versions of the
various tools, yadda yadda.