Add `configure' and compilation support for `clex'.
The important part of this commit adds support for configuring and
compiling the `clex' tool as part of the normal `configure'/`make'
dance.  See new files "clex/Makefile.am" and "clex/Makefile.in".

This commit also renames the primary `clex' source file from "c.lex" to
"clex.l".  Dot-el is what Automake expects.

This commit also tweaks "clex.l" so that it links under Mac OS X 10.8.
The source referred to a function `error()'; I changed these to call
`fprintf(stderr, ...)' instead.

The remaining changes come from the fact that I updated from Automake
version 1.13.1 to 1.13.2 on my laptop.  I'm sorry that these changes are
"mixed in" with the meaningful ones, but it was too much work for me to
try to separate them!