Don't short-circuit the execution of lex.
In a nutshell: Even when maintainer mode is off, run lex if the
generated C file seems to be out of date.  This is implemented by an
evil hack.  See the comments in "clex/Makefile.am" for more.

The rules for cleaning and distributing the lex-generated C files are
unchanged.  (The C file is not normally cleaned, and it is included in
distributions.)  We should probably change these behaviors, too, but
this commit does not make those changes.

This commit is mostly about avoiding developer aggravation.