Check the LLVM version at configure time.
The goal of this commit is to head off questions and/or confusion about
configure-time failures that arise from trying to use a too-old version
of LLVM.

Prior to this commit, AX_LLVM would (correctly) fail if one tried to use
a version of LLVM prior to 3.3.  The error message wasn't very specific,
however; it simply stated that the test program could not be compiled
and linked.  This commit makes a more specific error message possible.
Now, AX_LLVM can check for a minimum required LLVM version and issue an
error message if LLVM is too old.

In `m4/ax_llvm.m4':
  + Add minimum-version argument to AX_LLVM.
  + Be more careful about using AC_MSG_ERROR versus AC_MSG_FAILURE.

In `configure.ac':
  + Specify that LLVM 3.3 is the minimum required version.