Check for `clang-format' at configure time.
clang-format is a run-time dependency.  As such, we print a warning if
we can't find it at configure time, but we don't fail at configure time
if it is not found.

This commit searches for clang-format with the usual AC_PATH_PROG logic.
If --with-llvm is specified, should we look there for clang-format?
Should we look there preferentially?  This commit does neither of those.

This is basically a repeat of commit f352198 (Jun 28 2013), which was
undone by commit 5089b71 (Jul 1 2013).  It was undone in preparation for
the C-Reduce 2.1.0 release, which was targetting LLVM 3.3 release, which
has a version of clang-format that doesn't work so well for us.