Bind to external programs (e.g., `clang-format') at configure time.
At configure time, the `configure' script looks for programs that
C-Reduce will invoke at run time: i.e., `astyle', `clang-format',
`indent', and `topformflat'.

Previously, the `configure' script checked only for the benefit of
printing a useful warning about missing programs; the C-Reduce script
itself expected that the programs would be found again at run time in
the user's path.  Now, the configuration process remembers the pathnames
of the external programs that it finds.  C-Reduce uses those pathnames
to invoke the external programs at run time.

Program-not-found at configure time is still just a warning and causes a
(currently useless) relative filename to be configured into C-Reduce.
Program-not-found at run time causes C-Reduce to look in the user's
path.  (Maybe it should print a warning, too?  Future work if so.)

The primary usefulness of this change is to make it easier for one to
use the version of `clang-format' that comes with a particular copy of
LLVM.  If `--with-llvm' is specified at configure time, the configure
script looks there first for `clang-format'.  And obviously C-Reduce now
remembers what it found there --- no need to put the `--with-llvm'
directory into the path at run time.