Reimplement the "autoconfig hack" from commit 23121ed.
At commit 23121ed, Yang Chen made changes by hand to files `configure'
and `clang_delta/Makefile.in'.  Those changes were lost at commit
4178c1c, when I regenerated those files.

This commit restores the changes and does so in a way that allows them
to persist.

While experimenting with this, I realized that in LLVM 3.4 release,
`llvm-config --system-libs' is not implemented.  So this hack that Yang
implemented needs to be rethought.