Update the example for installing Clang/LLVM dependencies.
Thanks to Julien Cretin <creduce@ia0.eu> for discussion of what the
example command should actually be.

Julien notes that libtinfo-dev is actually a dependency of libedit-dev,
so one could simplify the command to just "apt-get install libedit-dev".
I didn't do this because I wanted the example to be more explicit.