Fixed potential SIGSEGV in ReplaceUndefinedFunction.cpp
Rewriting of functions is now restricted to locations directly in the
main file. The replacing functions are allowed to come from outside
of the main file. Only the replaced function and call sites have to
be in the main file.
Writing to other files is currently not supported. This fix
allows to have #include directives in the main file.