Be more aggressive about cleaning C++ compiler flags.
In particular, remove all warning-related options ("-W...").  While I
was in there, I tuned it up to also remove options of the form "-w",
"-g...", and "-O...".  We already removed "-pedantic...".

This change was instigated by the LLVM 3.7.0 binary distribution for
Ubuntu 14.04, available from
http://llvm.org/releases/3.7.0/clang+llvm-3.7.0-x86_64-linux-gnu-ubuntu-14.04.tar.xz

That binary distribution was compiled with the clang option
"-Wcovered-switch-default", which is not implemented by g++.

g++ is the compiler that our configure scripts choose by default for
building C-Reduce.  (This is standard behavior for Autoconf-based
projects.)

This means that if someone installs the precompiled LLVM 3.7.0
distribution on Ubuntu 14, the "out of the box build experience" for
C-Reduce fails at configure time, when we try to build the test LLVM
program (with g++).

No more trying to be clever about which -W options to keep and which to
discard!  Get rid of them all!