Use a better pipeline to process LLVM --cxxflags.
The 'sed' command in my previous commit is not portable.  At a minimum,
some sed's use "[[:<:]]" and "[[:>:]]" to match word boundaries, and
others use "\b".  In addition, the GNU Autoconf documentation says that
"^" is not portable.  Sigh.

So this commit replaces the 'sed' with a pipeline that uses 'grep' to do
the matching.  This is perhaps more portable, and bonus, it can be
written in a more human-readable way.  Huzzah!

The main trick is the new pipeline is to put the command-line options on
separate lines, using 'sed', so that we can use 'grep' to match the
options (lines) we want to remove.  Then the pipeline puts the
command-line options back onto a single line using 'xargs'.