Handle cases in which `bindir' references a symlink.
When we check if the `creduce' script is an installed copy, we obtain
the realpath of directory containing the script.

Previously we compared that realpath to the configured `bindir'.  That
test doesn't work right if `bindir' traverses a symbolic link.  Fix the
test by comparing to the realpath of `bindir'.

Thanks to Christian Schreiner for discovering this bug and submitting a
patch.  This commit is a tweaked version of Christian's patch.