Rewrite only the main source file
Previously, clang_delta tried to rewrite sources that were
not part of the main source file. This triggered a lot of crashes
when we process non-preprocessed code.

This patch teaches clang_delta to only rewrite the main source file.
The patch is based on the patches sent by Johan and Moritz.