Update tests to work with new C-Reduce protocol.
Until recently, C-Reduce would pass the name of the (one) file being
reduced on the command line to the "interestingness" script.  It no
longer does so.  The interestingness script needs to have built-in
knowledge of the file(s) being reduced.

This patch updates the tests to work with the current command-line
protocol.