Make clang_delta read include path and target triple from environment
When reducing files that are not fully pre-processed clang_delta need to
know where to find the unresolved include files, and when cross-reducing
files, i.e. when the reduced file is aimed at a different system than
the one where the reduction is running, the clang target triple needs to
be changed to get the correct type sizes.

This patch make clang_delta read the header search path from the content
of CREDUCE_INCLUDE_PATH and the target triple from the content of
CREDUCE_TARGET_TRIPLE.