Filter more problematic options out of `LLVMCXXFLAGS`.
Yang Chen reports that if you build Clang 3.8.0 with Clang, you end up
with `-fcolor-diagnostics` in the output of `llvm-config --cxxflags`.
This is a problem if you later try to build C-Reduce with GCC, because
GCC doesn't grok that flag.

With this commit, we now try to filter most diagnostics-reporting
options out of `llvm-config --cxxflags`.  There are a few that we still
do not match --- for example, Clang's `-fansi-escape-codes` and GCC's
`-fmessage-length` --- but I expect that these are rare enough that we
won't be troubled by them in practice.