Fix build w/Clang 3.8.0 binary package for OS X.
Fix the problem with building C-Reduce against the LLVM.org-provided
binary package of Clang 3.8.0, currently available at:
http://llvm.org/releases/3.8.0/clang+llvm-3.8.0-x86_64-apple-darwin.tar.xz

If you use that binary package, you'll find that the output of
`llvm-config --cxxflags` contains:

  -isysroot /Applications/Xcode.app/.../MacOSX10.10.sdk

This doesn't work with current versions of Xcode, at least not for John
Regehr and me, because recent versions of Xcode don't come with the OS X
10.10 SDK.

But no worries: we don't actually need those command-line options at all
in order to compile C-Reduce.  So now we strip them out.

The implementation of the stripping could be more precise.  Maybe we
should do it only atop OS X, and maybe we should look for the two
command-line words adjacent to each other, not separately.  But likely
the implementation in this commit is good enough for now, and probably
also likely good forever.