Use the git-version CMake module.
Use the newly added CMake module that determines the git hash of the
source repository.  Additionally, for the CMake-based build system,
(1) reproduce the logic from file `git-hash.sh' for determining the
appropriate git hash to build into C-Reduce; and (2) actually pass the
git hash to the C/C++ compilers.

This includes the ability to "burn the hash" into a release tarball.