Burn git hash into `clang_delta` via source file.
Generate a source file containing the current (short) git hash.  When
the git hash changes, regenerate that source file --- thus triggering
recompilation of that file and relinking of `clang_delta`.

Previously, we made the git hash available at compile time through a
command-line option to the compiler.  We did not have any kind of
dependency on the git hash value, however, so we couldn't trigger a
recompilation when it changed.  Worse: if it *did* change and we
noticed, we would in principle have needed to recompile every source
file.

The CMake-based build system already tracks the git hash as a
configuration dependency, and so it already suffers from the
recompile-everything problem described above.

This commit changes our strategy.  Now, we have an explicit dependency
on the git hash value, and when it changes, we only need to regenerate
and recompile one source file.  This is "sexy," but more significantly,
it will help us to make the CMake-based build system more usable for
developers.  (This commit does not update the CMake-based build system,
however.  Forthcoming!)