enable multiple-rewrites for remove-unused-var pass
There are still some issues with group-decls when multiple-rewrites are
enabled, e.g., the pass will generate bad code for "int i, j" if both
i and j are unused. However, I think in practice it may not be a big
issue because when the pass is used as a binsrch-clang-pass from creduce,
it shouldn't be worse than removing one unused var at a time.