Enable assertions by default in CMake build path.
This behavior is consistent with the Automake-based build system.

Update the `INSTALL` documentation to match the current default
behavior.  Note that the description of the default for the
Automake-based build system was wrong.  I'm going to say that is what
confused me when I implemented the CMake-based build path :-).

Thanks to @oriceemple at GitHub for reporting this issue.

This resolves #116.