Add comment about `CLANG_INCLUDE_DIRS`.
The variable `CLANG_INCLUDE_DIRS` was added in Clang r294207, which
is after the LLVM 4.0.0 release.  Keep it here now, because it is
useful to people who build C-Reduce atop the development versions of
LLVM and Clang.  In the future, when C-Reduce requires LLVM 5, we can
just remove this comment.