Add initial implementation of `--with-clang` configure option.
Use `--with-clang` to identify a separate Clang development tree where
the Automake-based build process should find Clang headers and
libraries.  This is useful if you compile LLVM and Clang from source
and keep them in separate build trees.

Moritz Pflanzer implemented support for separate LLVM and Clang trees
in commit fea887e, for the CMake-based build process only.  This
commit implements similar support for the Automake-based build process.

This commit addresses issue #133 and is related to pull request #132.

I'm not entirely happy that `clang_delta/Makefile.am` handles the LLVM
and Clang configuration options differently.  For LLVM, the Makefile
queries `llvm-config` when Make is run; but for this new Clang stuff,
we burn the Clang-related stuff into the Makefile when `configure` is
run.  Maybe I will fix this; maybe I won't.

Also: Even when `--with-clang` is specified, the Makefile still causes
the build to look in directory `$(LLVMINCLUDEDIR)/clang` for Clang
headers (in addition to any that `configure` found because
`--with-clang` was specified).  Maybe I will fix this; maybe I won't.