Remove doc about using separate LLVM and Clang build trees.
The described feature doesn't work with the LLVM/Clang 4.0 release;
see the discussion for pull request #132.  (The underlying support in
Clang's CMake-based build system was added after the 4.0 release.)

I'm aiming to make a release of C-Reduce that is compatible with the
LLVM 4.0 release, so I don't want to describe a feature that doesn't
work with that.

I'm going to leave the implementation of the feature in place for the
release, however (i.e., support for `configure --with-clang` and
`cmake Clang_DIR=...`).  It will just be undocumented!

The documentation will come right back after I make the upcoming
C-Reduce release.