Handle anonymous unions are declared with other decls
In some rare cases, anonymous unions can be declared along with
other declarations such as FunctionDecl. In these cases, we have
to add new DDSet while processing the declarator. Furthermore,
we don't need to insert the declarator into the DDSet, because
the union keyword will be replaced by the routine for processing
the RecordDecl.