Update info about dependencis of pre-build binary packages.
Pick a more recent OS as the example.  Update the wording to match the
current LLVM downloads website.  Mention that the Ubuntu 18.04 pre-built
binaries package also depends on `libxml2`.

(Actually, when using the Ubuntu 18.04 pre-built binaries package, the
current CMake-based build path *does not* require `libxml2`, because we
don't link against the components of LLVM that depend on that library.
The Autoconf-based build path is less precise in this regard.  It uses
`llvm-config --system-libs` to get the system libraries, and that answer
contains "-lxml2".  Whatever.  In the INSTALL instructions, just tell
people what kind of issues they might encounter.)