Largely revert commit be741fc.
Go back to the old way of computing LLVM_LIBS and then not using it :-).
The "new way" --- namely, invoking `llvm-config` --- might be
resurrected and completed in the future.  For now, toward getting a new
release out the door, I'm going back to the old way.  (It has the minor
advantage of having one less CMake-time dependency.)