On the CMake-based build path, set rpath for `clang_delta`.
Remember where the LLVM libraries are by specifying the rpath for
`clang_delta`.  This is important when the LLVM libraries are shared
libraries.

The Automake-based build path is still broken in this regard.  Until I
fix it (if I don't obsolete the Automake-based build path first!), users
will have to do something like setting LD_LIBRARY_PATH, or installing
LLVM in a well-known place, or somesuch.