Set `BUILD_WITH_INSTALL_RPATH` property on target `clang_delta`.
Set the `BUILD_WITH_INSTALL_RPATH` property, and do not set the
`BUILD_RPATH` property.  The effect is the same, but this method works
with older versions of CMake, e.g., CMake 3.5.1, which is the version
you get if you `apt-get install cmake` on Ubuntu 16.04.

`BUILD_WITH_INSTALL_RPATH` is in the CMake docs back to at least CMake
2.8.0.  `BUILD_RPATH` was added in CMake 3.8.